+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
     an array of `[styles]`
 
 ### Fixed
+- `textfield`
+  - Fix `firstUpdated` code path to avoid a race condition in initialization with `notched-outline`
 
 - An infinite loop caused by having both `_getUpdateComplete` and
   `getUpdateComplete` introduced by lit-element 2.5.0